% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getCompany.R
\name{getCompany}
\alias{getCompany}
\title{Company Profile API}
\usage{
getCompany(token, universal_name = NULL, email_domain = NULL,
  company_id = NULL)
}
\arguments{
\item{token}{Authorization token}

\item{universal_name}{LinkedIn universal company name}

\item{email_domain}{Email domain used}

\item{company_id}{LinkedIn company ID}
}
\value{
Returns company profile data, such as LinkedIn ID, name, universal-name, email-domains, company-type, ticker, website-url, industries, status, twitter handle, employee-count-range, specialties, locations, description, founded-year, and number of followers.
}
\description{
The Company Profile API retrieves and displays company profiles using a company ID, a universal name, or an email domain.
}
\details{
Universal name needs to be the exact name seen at the end of the URL on the company page in Linkedin
}
\examples{
\dontrun{

company.name <- getCompany(token=in.auth, universal_name="Facebook")

company.email <- getCompany(token=in.auth, email_domain = "columbia.edu")

# Main Columbia Name:
company.id <- getCompany(token=in.auth, company_id = company.email$company_id[14])


}
}

