% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getGroupPosts.R
\name{getGroupPosts}
\alias{getGroupPosts}
\title{Get Posts from your LinkedIn Groups}
\usage{
getGroupPosts(token)
}
\arguments{
\item{token}{Authorization token}
}
\value{
Returns posts of groups you belong to
}
\description{
Currently this will retreive only the last 10 posts from each group
}
\examples{
\dontrun{

my.groups <- getGroupPosts(in.auth)
}
}

