% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getGroups.R
\name{getGroups}
\alias{getGroups}
\title{Get info from your LinkedIn Groups}
\usage{
getGroups(token, details = FALSE)
}
\arguments{
\item{token}{Authorization token}

\item{details}{Group details, TRUE or FALSE}
}
\value{
Returns information about what groups you belong to, either with or without group details.
}
\description{
Get info from your LinkedIn Groups
}
\examples{
\dontrun{

my.groups <- getGroups(in.auth, details=TRUE)
}
}

