% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getJobs.R
\name{getJobs}
\alias{getJobs}
\title{LinkedIn Bookmarked and Recommended Jobs}
\usage{
getJobs(token, suggestions = TRUE, bookmarks = FALSE)
}
\arguments{
\item{token}{Authorization token}

\item{suggestions}{LinkedIn job recommendations}

\item{bookmarks}{Jobs you've bookmarked}
}
\value{
Returns LinkedIn recommended or bookmarked jobs
}
\description{
LinkedIn Bookmarked and Recommended Jobs
}
\examples{
\dontrun{

job.suggestions <- getJobs(in.auth, suggestions=TRUE)
job.bookmarks <- getJobs(in.auth, bookmarks=TRUE)

}
}

