% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getMyConnections.R
\name{getMyConenctions}
\alias{getMyConenctions}
\title{Get LinkedIn Connections}
\usage{
getMyConenctions(token)
}
\arguments{
\item{token}{Authorization token}
}
\value{
Returns a dataframe of LinkedIn connections
}
\description{
Get LinkedIn Connections
}
\examples{
\dontrun{

my.connections <- getMyConnections(in.auth)
}
}

