% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getProfile.R
\name{getProfile}
\alias{getProfile}
\title{Get LinkedIn Profile Information}
\usage{
getProfile(token, connections = FALSE, id = NULL)
}
\arguments{
\item{token}{Authorization token}

\item{connections}{TRUE or FALSE. If TRUE, will return profile information of your connections}

\item{id}{Numeric ID number of a LinkedIn member}
}
\value{
Returns profile information of yourself, your connections, or an idividual LinkedIn member
}
\description{
Get LinkedIn Profile Information
}
\examples{
\dontrun{

profiles <- getProfile(in.auth, connections=TRUE)
}
}

