% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/Rlinkedin-package.R
\docType{package}
\name{Rlinkedin-package}
\alias{Rlinkedin}
\alias{Rlinkedin-package}
\title{Access to LinkedIn API via R}
\description{
This is an R package that provides a series of functions that allow users
to access the LinkedIn API to get information about connections, search for people,
search for jobs, share updates with your network, and create group discussions.
}
\author{
Michael Piccirilli \email{mrp2181@columbia.edu}
}
\seealso{
\code{\link{inOAuth}}, \code{\link{getProfile}},
\code{\link{getMyConnections}}, \code{\link{getGroupPosts}},
\code{\link{getGroups}}, \code{\link{getJobs}},
\code{\link{searchJobs}}, \code{\link{searchPeople}},
\code{\link{submitGroupPost}}, \code{\link{submitShare}},
\code{\link{getCompany}}, \code{\link{searchCompanies}}
}

