% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/searchCompanies.R
\name{searchCompanies}
\alias{searchCompanies}
\title{Company Search API: Search for Companies on LinkedIn}
\usage{
searchCompanies(token, keywords)
}
\arguments{
\item{token}{Authorization token}

\item{keywords}{Keyword people search}
}
\value{
Returns a search of companies based on keywords, industries, or location
}
\description{
The Company Search API enables search across company pages.
}
\examples{
\dontrun{

search.comp <- searchCompanies(in.auth, "LinkedIn")

}
}

