% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/searchCompany.R
\name{searchCompany}
\alias{searchCompany}
\title{Company Profile API: Search for Companies on LinkedIn}
\usage{
searchCompany(token, universal_name = NULL, email_domain = NULL,
  company_id = NULL)
}
\arguments{
\item{token}{Authorization token}

\item{universal_name}{LinkedIn universal company name}

\item{email_domain}{Email domain used}

\item{company_id}{LinkedIn company ID}
}
\value{
Returns company profile data, such as LinkedIn ID, name, universal-name, email-domains, company-type, ticker, website-url, industries, status, twitter handle, employee-count-range, specialties, locations, description, founded-year, and number of followers.
}
\description{
The Company Profile API retrieves and displays company profiles based on the company ID or universal name.
}
\details{
Universal name needs to be the exact name seen at the end of the URL on the company page in Linkedin
}
\examples{
\dontrun{

company.name <- searchCompany(token=in.auth, universal_name="Facebook")

company.email <- searchCompany(token=in.auth, email_domain = "columbia.edu")

# Main Columbia Name:
company.id <- searchCompany(token=in.auth, company_id = company.email$company_id[14])


}
}

