% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/searchJobs.R
\name{searchJobs}
\alias{searchJobs}
\title{Job Search API: Search for Jobs on LinkedIn}
\usage{
searchJobs(token, keywords, count = 20)
}
\arguments{
\item{token}{Authorization token}

\item{count}{Authorization token}

\item{keyword}{Keyword job search}
}
\value{
Returns jobs based input parameters
}
\description{
Job Search API is a part of our Vetted API Access Program. You must apply and get LinkedIn's approval before using this API.
}
\examples{
\dontrun{

search.results <- searchJobs(in.auth, "data scientist")
}
}

