% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/searchPeople.R
\name{searchPeople}
\alias{searchPeople}
\title{People Search API: Search for People on LinkedIn}
\usage{
searchPeople(token, keywords = NULL, first_name = NULL, last_name = NULL,
  company_name = NULL, current_company = NULL, title = NULL,
  current_title = NULL, school_name = NULL, current_school = NULL,
  country_code = NULL, postal_code = NULL, distance = NULL)
}
\arguments{
\item{token}{Authorization token}

\item{keywords}{Keyword people search}

\item{first_name}{First Name}

\item{last_name}{Last Name}

\item{company_name}{Company Name}

\item{current_company}{TRUE/FALSE}

\item{title}{Job Title}

\item{current_title}{TRUE/FALSE}

\item{school_name}{Name of a school}

\item{current_school}{TRUE/FALSE}

\item{country_code}{TRUE/FALSE}

\item{postal_code}{Must be combined with the country-code parameter.}

\item{distance}{Distance Matches members within a distance from a central point. This is measured in miles.}
}
\value{
Returns people search based input parameters
}
\description{
The People Search API returns information about people. It lets you implement most of what shows up when you do a search for "People" in the top right box on LinkedIn.com.
}
\details{
People Search API is a part of LinkedIn's Vetted API Access Program. You must apply and get LinkedIn's approval before using this API.

Up to 100 returns per search, 10 returns per page. Each page is one API call.
}
\examples{
\dontrun{

search.results <- searchPeople(token=in.auth, first_name="Michael", last_name="Piccirilli")
}
}

