% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/submitGroupPost.R
\name{submitGroupPost}
\alias{submitGroupPost}
\title{Create a Group Discussion Post}
\usage{
submitGroupPost(token, group_id, disc_title = NULL, disc_summary = NULL,
  content_url = NULL, content_img = NULL, content_title = NULL,
  content_desc = NULL)
}
\arguments{
\item{token}{Authorization token}

\item{group_id}{Numeric Group ID}

\item{disc_title}{Group Discussion Title, required}

\item{disc_summary}{Group Discussion Summary, required}

\item{content_url}{Url for content, optinal}

\item{content_img}{Url for an image, optional}

\item{content_title}{Title for content, required}

\item{content_desc}{Description of content, optional}
}
\value{
Creates a group discussion
}
\description{
Must include a minimum of a discussion title, discussion summary, and content title"
}
\examples{
\dontrun{

my.groups <- getGroups(in.auth)

id <- my.groups$group_id[1]
disc.title <- "Test connecting to the LinkedIn API via R"
disc.summary <- "Im creating an R package to connect to the LinkedIn API, this is a test post from R!"
url <- "https://github.com/mpiccirilli"
content.desc <- "Dev version of access to LinkedIn API via R. Collaboration is welcomed!"

submitGroupPost(in.auth, group_id=id, disc_title=disc.title, disc_summary=disc.summary, content_url=url, content_desc=content.desc)
}
}

