% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/submitShare.R
\name{submitShare}
\alias{submitShare}
\title{Share a Network Update}
\usage{
submitShare(token, comment = NULL, content_title = NULL,
  content_desc = NULL, content_url = NULL, content_img = NULL,
  visibility = "anyone")
}
\arguments{
\item{token}{Authorization token}

\item{comment}{Headline of the post}

\item{content_title}{Title of the post}

\item{content_desc}{Description of the post}

\item{content_url}{Url to content}

\item{content_img}{Url to image}

\item{visibility}{'anyone' or 'connections-only'}
}
\value{
Creates a post to the network feed
}
\description{
If one of the content elements is specified you must also include include a url for the post
}
\examples{
\dontrun{

comment <- "Test connecting to the LinkedIn API via R"
title <- "Im creating an R package to connect to the LinkedIn API, this is a test post from R!"
url <- "https://github.com/mpiccirilli"
desc <- "Dev version of access to LinkedIn API via R. Collaboration is welcomed!"

submitShare(in.auth, comment=comment, content_tile=title, content_url=url, content_desc=desc)
}
}

