% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getGroupPosts.R
\name{getGroupPosts}
\alias{getGroupPosts}
\title{Extract Posts from your LinkedIn Groups}
\usage{
getGroupPosts(token)
}
\arguments{
\item{token}{Authorization token.}
}
\value{
Returns a dataframe with the 10 most recent posts from each
group you belong to.
}
\description{
\code{getGroupPosts} will retrieve posts from each LinkedIn group you belong to.
}
\examples{
\dontrun{

my.groups <- getGroupPosts(in.auth)
}
}
\author{
Michael Piccirilli \email{michael.r.piccirilli@gmail.com}
}
\seealso{
\code{\link{getGroups}} \code{\link{submitGroupPost}}
}

