% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getJobs.R
\name{getJobs}
\alias{getJobs}
\title{Bookmarked and Recommended Jobs on LinkedIn}
\usage{
getJobs(token, suggestions = NULL, bookmarks = NULL)
}
\arguments{
\item{token}{Authorization token}

\item{suggestions}{Suggestions LinkedIn job recommendations}

\item{bookmarks}{Bookmarks LinkedIn jobs you've bookmarked}
}
\value{
Returns LinkedIn recommended or bookmarked jobs
}
\description{
Bookmarked and Recommended Jobs on LinkedIn
}
\examples{
\dontrun{

job.suggestions <- getJobs(in.auth, suggestions=TRUE)
job.bookmarks <- getJobs(in.auth, bookmarks=TRUE)

job.fail <- getJobs(in.auth) \\# Will return null

}
}

