% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getJobs.R
\name{getJobs}
\alias{getJobs}
\title{LinkedIn Job Recommendations and Job Search}
\usage{
getJobs(token, suggestions = TRUE)
}
\arguments{
\item{token}{Authorization token}

\item{suggestions}{LinkedIn job recommendations}
}
\value{
Returns LinkedIn job recommendation or search for jobs via LinkedIn
}
\description{
LinkedIn Job Recommendations and Job Search
}
\examples{
\dontrun{

my.jobs <- getJobs(in.auth, suggestions=TRUE)
}
}

