% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/inOAuth.R
\name{inOAuth}
\alias{inOAuth}
\title{LinkedIn Job Recommendations and Job Search}
\usage{
inOAuth(application_name, consumer_key, consumer_secret)
}
\arguments{
\item{application_name}{Name of your application}

\item{consumer_key}{Consumer API Key of your application}

\item{consumer_secret}{Consumer Secret Key of your application}
}
\value{
Authorization token to be used in other functions
}
\description{
Establishes a connection to the LinkedIn API.
}
\examples{
\dontrun{

in.auth <- inOAuth("myappname", "your_consumer_key", "your_consumer_secret")
}
}

