% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/submitGroupPost.R
\name{submitGroupPost}
\alias{submitGroupPost}
\title{Create a Group Discussion Post}
\usage{
submitGroupPost(token, group_id, disc_title = NULL, disc_summary = NULL,
  content_url = NULL, img_url = NULL, content_title = NULL,
  content_desc = NULL)
}
\arguments{
\item{token}{Authorization token}

\item{group_id}{Numeric Group ID}

\item{disc_title}{Group Discussion Title}

\item{disc_summary}{Group Discussion Summary}

\item{content_url}{Url for content, optinal}

\item{img_url}{Url for an image, optional}

\item{content_title}{Title for content, optional}

\item{content_desc}{Description of content, optional}
}
\value{
Creates a group discussion
}
\description{
Create a Group Discussion Post
}
\examples{
\dontrun{

id <- 77616  \\# id number for The R Project for Statistical Computing
disc.title <- "Test connecting to the LinkedIn API via R"
disc.summary <- "Im creating an R package to connect to the LinkedIn API, this is a test post from R!"
url <- "https://github.com/mpiccirilli"
content.desc <- "Dev version of access to LinkedIn API via R. Collaboration is welcomed!"

submitGroupPost(in.auth, 77616, disc_title=disc.title, disc_summary=disc.summary, content_url=url, content_desc=content.desc)
}
}

